etl/requirements.txt
pydantic==1.9.0
isort==5.10.1
python-dotenv==0.19.2
flake8==4.0.1
psycopg2-binary==2.9.3
elasticsearch==7.17.1